# [WIP] Migrate `glu` & `glu_backward` from the TH to Aten (CUDA)