--- a/BUILD.bazel
+++ b/BUILD.bazel
@@ -437,7 +437,6 @@ filegroup(
         "aten/src/THCUNN/BCECriterion.cu.cc",
         "aten/src/THCUNN/ClassNLLCriterion.cu.cc",
         "aten/src/THCUNN/ELU.cu.cc",
-        "aten/src/THCUNN/GatedLinearUnit.cu.cc",
         "aten/src/THCUNN/HardTanh.cu.cc",
         "aten/src/THCUNN/LeakyReLU.cu.cc",
         "aten/src/THCUNN/LogSigmoid.cu.cc",